Replaced return values and parameters with suitable NS(U)Integer to avoid long->int issues in 64 bit. Appropriate cast where warranted.

+++ Other Sources/Foreign Code/SubviewTableView/SubviewTableViewController.h
@@ -81,6 +81,6 @@ have otherwise used "reloadData" from NSTableView:
 // The view retreived will not be retained, and will be resized to fit the
 // cell in the table view. Please adjust the row height and column width in
 // ib (or in code) to make sure that it is appropriate for the views used.
-- (NSView *) tableView:(NSTableView *) tableView viewForRow:(int) row;
+- (NSView *) tableView:(NSTableView *) tableView viewForRow:(NSInteger) row;
 
 @end
