+++ Pixen Application/Tools/Tool Palette/PXToolPaletteController.m
@@ -208,7 +208,7 @@
 	if (_locked) {
 		return usingRightToolBeforeLock;
 	}
-	int modFlags = [[[NSApplication sharedApplication] currentEvent] modifierFlags];
+	NSUInteger modFlags = [[[NSApplication sharedApplication] currentEvent] modifierFlags];
 	BOOL controlReallyDown = ((modFlags & NSControlKeyMask) == NSControlKeyMask); // sometimes we don't actually receive the message.  what a bother...
 	if (controlKeyDown != controlReallyDown) {
 		controlKeyDown = controlReallyDown;
@@ -253,12 +253,12 @@
 		[leftSwitcher keyDown:event fromCanvasController:cc];
 }
 
-- (BOOL)keyWasDown:(unsigned int)mask
+- (BOOL)keyWasDown:(NSUInteger)mask
 {
     return (keyMask & mask) == mask;
 }
 
-- (BOOL)isMask:(unsigned int)newMask upEventForModifierMask:(unsigned int)mask
+- (BOOL)isMask:(NSUInteger)newMask upEventForModifierMask:(unsigned int)mask
 {
     return [self keyWasDown:mask] && ((newMask & mask) == 0x0000);
 }
